partial changes work in progress

+++ Falcon.Service/Master/MasterService.cs
@@ -37,32 +37,32 @@ namespace Falcon.Service.MasterRepository
             model.sessionList.AddRange(result.Tables["SessionMaster"].AsEnumerable()
                                     .Select(row => new SessionModel()
                                     {
-                                        myId = Convert.ToInt32(row.Field<decimal>("myId")),
+                                        myId = Convert.ToInt32(row.Field<int>("myId")),
                                         Name = row.Field<string>("SessionName")
                                     }));
 
             model.classList.AddRange(result.Tables["ClassMaster"].AsEnumerable()
                                     .Select(row => new ClassModel()
                                     {
-                                        myId = Convert.ToInt32(row.Field<decimal>("myId")),
+                                        myId = Convert.ToInt32(row.Field<int>("myId")),
                                         Name = row.Field<string>("ClassName")
                                     }));
 
             model.sectionList.AddRange(result.Tables["SectionMaster"].AsEnumerable()
                                     .Select(row => new SectionModel()
                                     {
-                                        myId = Convert.ToInt32(row.Field<decimal>("myId")),
+                                        myId = Convert.ToInt32(row.Field<int>("myId")),
                                         Name = row.Field<string>("SectionCode")
                                     }));
 
             model.ClassXrefList.AddRange(result.Tables["ClassXref"].AsEnumerable()
                                     .Select(row => new ClassXref()
                                     {
-                                        myId = Convert.ToInt32(row.Field<decimal>("myId")),
-                                        ClassId = Convert.ToInt32(row.Field<decimal>("refClassId")),
-                                        SectionId = Convert.ToInt32(row.Field<decimal>("refSecId")),
-                                        SessionId = Convert.ToInt32(row.Field<decimal>("refSessionId")),
-                                        Strength = 0/*Convert.ToInt32(row.Field<decimal>("Strength"))*/
+                                        myId = Convert.ToInt32(row.Field<int>("myId")),
+                                        ClassId = Convert.ToInt32(row.Field<int>("refClassId")),
+                                        SectionId = Convert.ToInt32(row.Field<int>("refSecId")),
+                                        SessionId = Convert.ToInt32(row.Field<int>("refSessionId")),
+                                        Strength = 0/*Convert.ToInt32(row.Field<int>("Strength"))*/
                                     }));
 
             return model;
@@ -71,10 +71,10 @@ namespace Falcon.Service.MasterRepository
         public bool UpdateClassConfig(string[] AllKeys)
         {
             DataTable classXrefDataTable = new DataTable("ClassXref");
-            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_CheckedId", DataType = typeof(decimal) });
-            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_ClassID", DataType = typeof(decimal) });
-            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_SectionId", DataType = typeof(decimal) });
-            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_SessionID", DataType = typeof(decimal) });
+            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_CheckedId", DataType = typeof(int) });
+            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_ClassID", DataType = typeof(int) });
+            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_SectionId", DataType = typeof(int) });
+            classXrefDataTable.Columns.Add(new DataColumn { ColumnName = "UT_SessionID", DataType = typeof(int) });
 
             DataRow row;
 
@@ -85,10 +85,10 @@ namespace Falcon.Service.MasterRepository
                 if (classArray[0] == "0")
                 {
                     row = classXrefDataTable.NewRow();
-                    row["UT_CheckedId"] = Convert.ToDecimal(classArray[0]);
-                    row["UT_ClassID"] = Convert.ToDecimal(classArray[2]);
-                    row["UT_SectionId"] = Convert.ToDecimal(classArray[3]);
-                    row["UT_SessionID"] = Convert.ToDecimal(classArray[1]);
+                    row["UT_CheckedId"] = Convert.ToInt32(classArray[0]);
+                    row["UT_ClassID"] = Convert.ToInt32(classArray[2]);
+                    row["UT_SectionId"] = Convert.ToInt32(classArray[3]);
+                    row["UT_SessionID"] = Convert.ToInt32(classArray[1]);
 
                     classXrefDataTable.Rows.Add(row);
                 }
